Add tests to validate the handling of new lines.
